<feat> 变长数据架构实现

+++ src/observer/sql/parser/parse_defs.h
@@ -165,7 +165,7 @@ struct AttrInfoSqlNode
 {
   AttrType    type;       ///< Type of attribute
   std::string name;       ///< Attribute name
-  size_t      length;     ///< Length of attribute
+  int      length;     ///< Length of attribute
 };
 
 /**
